Fix readme so that users can set configuration at compile time
We need to allow for dynamic configuration at runtime and make use of
command line args or environment variables. Also, we can pass
configurations from the scheduler to the executors at runtime as
needed. This is part of our roadmap. Meanwhile, users often encounter
problems if their executors are not bundled with the proper
mesos-site.xml.