Enable Framework Authentication with CRAM-MD5
This has been tested on DCOS clusters.  The Config changes refer to
configuration which does not exist in this change.  It has been tested
with that configuration present and absent, and works as expected.  The
framework fails authentication with incorrect credentials, succeeds with
correct credentials, and succeeds in registering without authentication
when the config is absent.

See this pull request for the changes required to make SASL
authentication operable in DCOS.

https://github.com/mesosphere/dcos-image/pull/74